backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-celery-beat>=2.5
django-celery-results>=2.5
celery>=5.3
redis>=5.0
psycopg2-binary>=2.9
Pillow>=10.0
python-dotenv>=1.0
openai>=1.0
gunicorn>=21.0
pinecone
firebase-admin>=6.0.0
exponent_server_sdk>=1.0.0
charset-normalizer>=3.0
